pyproject.toml
[build-system]
requires = ["setuptools>=68.2.2", "wheel>=0.41.2", "setuptools_scm>=8.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "proventra_core"
version = "0.0.1"
description = "Prompt injection detection and prevention library for LLM applications"
readme = "README.md"
requires-python = ">=3.11,<4.0"
license = {text = "MIT"}
authors = [
    {name = "Proventra Team"}
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "torch>=2.2.1",
    "transformers>=4.38.2",
    "numpy>=1.26.4",
    "langchain>=0.1.4",
    "langchain-core>=0.1.27",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.5.2"
]

[project.optional-dependencies]
# LLM providers
gemini = ["langchain-google-genai>=0.0.10"]
anthropic = ["langchain-anthropic>=0.1.3"]
openai = ["langchain-openai>=0.0.8"]
mistral = ["langchain-mistralai>=0.0.5"]
all = [
    "langchain-google-genai>=0.0.10",
    "langchain-anthropic>=0.1.3", 
    "langchain-openai>=0.0.8",
    "langchain-mistralai>=0.0.5"
]

# Deployment extras
api = ["fastapi>=0.109.2", "uvicorn>=0.27.1"]
runpod = ["runpod>=1.5.0"]

# Development tools
dev = [
    "pytest>=8.0.2", 
    "isort>=5.13.2", 
    "mypy>=1.7.0",
    "ruff>=0.2.2"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

# Enable Ruff formatter
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

# Ignore missing imports for specific external packages
[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "torch.*",
    "numpy.*",
    "tiktoken.*",
]
ignore_missing_imports = true 